execution/discord_bot/requirements.txt
discord.py>=2.3.0
anthropic>=0.40.0
aiohttp>=3.9.0
requests>=2.31.0
